Add keys 'UsePlanPsector' and 'macroExploitation' to enhance control of grafanalytics function

+++ ws/fct/ws_gw_fct_grafanalytics_mapzones.sql
@@ -37,28 +37,28 @@ update connec set sector_id=0, dma_id=0, dqa_id=0, presszone_id=0
 ----------
 TO EXECUTE
 ----------
+select * from SCHEMA_NAME.exploitation
+-- QUERY SAMPLE
+SELECT gw_fct_grafanalytics_mapzones('{"data":{"parameters":{"grafClass":"DMA", "exploitation": "[86]", "checkData":false, "updateFeature":"TRUE", 
+"updateMapZone":2, "geomParamUpdate":15,"debug":"false", "usePlanPsector":false, "forceOpen":[1,2,3], "forceClosed":"[2,3,4]"}}}');
 
--- SECTOR
-SELECT SCHEMA_NAME.gw_fct_grafanalytics_mapzones('{"data":{"parameters":{"grafClass":"SECTOR", "exploitation": "[1,2]", "checkData": false, "updateFeature":"FALSE", "updateMapZone":2, "geomParamUpdate":15, "debug":"false"}}}');
+--SECTOR
 SELECT gw_fct_grafanalytics_mapzones('{"data":{"parameters":{"grafClass":"SECTOR", "node":"113952", "updateFeature":TRUE}}}');
 SELECT count(*), log_message FROM audit_log_data WHERE fid=130 AND cur_user=current_user group by log_message order by 2 --SECTOR
 SELECT sector_id, count(sector_id) from v_edit_arc group by sector_id order by 1;
 
-
 -- DMA
-SELECT gw_fct_grafanalytics_mapzones('{"data":{"parameters":{"grafClass":"DMA", "exploitation": "[1]", "checkData": false,"updateFeature":"TRUE", "updateMapZone":2, "geomParamUpdate":15,"debug":"false"}}}');
+SELECT gw_fct_grafanalytics_mapzones('{"data":{"parameters":{"grafClass":"DMA", "exploitation": "[1]", "checkData": false,"updateFeature":"TRUE", "updateMapZone":2, "geomParamUpdate":15,"debug":"false", "forceOpen":[1,2,3], "forceClose":"[2,3,4]"}}}');
 SELECT gw_fct_grafanalytics_mapzones('{"data":{"parameters":{"grafClass":"DMA", "node":"1046", "updateFeature":"TRUE", "updateMapZone":2,"concaveHullParam":0.85,"debug":"false"}}}');
 SELECT count(*), log_message FROM audit_log_data WHERE fid=145 AND cur_user=current_user group by log_message order by 2 --DMA
 SELECT dma_id, count(dma_id) from v_edit_arc  group by dma_id order by 1;
 UPDATE arc SET dma_id=0
-
 
 -- DQA
 SELECT gw_fct_grafanalytics_mapzones('{"data":{"parameters":{"grafClass":"DQA", "exploitation": "[1,2]", "checkData": false,"updateFeature":"TRUE", "updateMapZone":2 , "geomParamUpdate":15, "debug":"false"}}}');
 SELECT gw_fct_grafanalytics_mapzones('{"data":{"parameters":{"grafClass":"DQA", "node":"113952", "updateFeature":TRUE}}}');
 SELECT count(*), log_message FROM audit_log_data WHERE fid=144 AND cur_user=current_user group by log_message order by 2 --DQA
 SELECT dqa_id, count(dma_id) from v_edit_arc  group by dqa_id order by 1;
-
 
 -- PRESZZONE
 SELECT gw_fct_grafanalytics_mapzones('{"data":{"parameters":{"grafClass":"PRESSZONE","exploitation":"[1]", "checkData": false, "updateFeature":"TRUE", "updateMapZone":2, "geomParamUpdate":15,"debug":"false"}}}');
@@ -110,6 +110,7 @@ v_cont1 integer default 0;
 v_class text;
 v_feature record;
 v_expl json;
+v_macroexpl json;
 v_data json;
 v_fid integer;
 v_nodeid text;
@@ -146,6 +147,8 @@ v_status text;
 v_message text;
 v_checkdata boolean;
 v_mapzonename text;
+v_parameters json;
+v_usepsector boolean;
 
 BEGIN
 	-- Search path
@@ -164,8 +167,11 @@ BEGIN
 	v_updatemapzgeom = (SELECT ((p_data::json->>'data')::json->>'parameters')::json->>'updateMapZone');
 	v_geomparamupdate = (SELECT ((p_data::json->>'data')::json->>'parameters')::json->>'geomParamUpdate');
 	v_expl = (SELECT ((p_data::json->>'data')::json->>'parameters')::json->>'exploitation');
+	v_macroexpl = (SELECT ((p_data::json->>'data')::json->>'parameters')::json->>'macroExploitation');
+	v_usepsector = (SELECT ((p_data::json->>'data')::json->>'parameters')::json->>'usePlanPsector');
 	v_debug = (SELECT ((p_data::json->>'data')::json->>'parameters')::json->>'debug');
 	v_checkdata = (SELECT ((p_data::json->>'data')::json->>'parameters')::json->>'checkData');
+	v_parameters = (SELECT ((p_data::json->>'data')::json->>'parameters'));
 
 	-- select config values
 	SELECT giswater, epsg INTO v_version, v_srid FROM sys_version order by 1 desc limit 1;
@@ -262,13 +268,29 @@ BEGIN
 		-- reset selectors
 		DELETE FROM selector_state WHERE cur_user=current_user;
 		INSERT INTO selector_state (state_id, cur_user) VALUES (1, current_user);
-		--DELETE FROM selector_psector WHERE cur_user=current_user;
 			
 		-- reset exploitation
 		IF v_expl IS NOT NULL THEN
 			DELETE FROM selector_expl WHERE cur_user=current_user;
 			INSERT INTO selector_expl (expl_id, cur_user) 
 			SELECT expl_id, current_user FROM exploitation WHERE expl_id IN	(SELECT (json_array_elements_text(v_expl))::integer);
+		END IF;
+
+		IF v_macroexpl IS NOT NULL THEN
+			DELETE FROM selector_expl WHERE cur_user=current_user;
+			INSERT INTO selector_expl (expl_id, cur_user) 
+			SELECT expl_id, current_user FROM exploitation WHERE macroexpl_id IN (SELECT (json_array_elements_text(v_macroexpl))::integer);
+		END IF;
+
+		IF v_usepsector IS NOT TRUE THEN
+		
+			-- save psector selector 
+			DELETE FROM temp_table WHERE fid=199 AND cur_user=current_user;
+			INSERT INTO temp_table (fid, text_column)
+			SELECT 199, (array_agg(psector_id)) FROM selector_psector WHERE cur_user=current_user;
+
+			-- set psector selector
+			DELETE FROM selector_psector WHERE cur_user=current_user;
 		END IF;
 
 		-- start build log message
@@ -314,15 +336,25 @@ BEGIN
 			v_querytext  = 'UPDATE temp_anlgraf SET flag=1 WHERE 
 					node_1::integer IN('||v_text||' UNION
 					SELECT (a.node_id::integer) FROM node a JOIN cat_node b ON nodecat_id=b.id JOIN cat_feature_node c ON c.id=b.nodetype_id 
-					LEFT JOIN man_valve d ON a.node_id::integer=d.node_id::integer JOIN temp_anlgraf e ON a.node_id::integer=e.node_1::integer WHERE closed=TRUE)
+					LEFT JOIN man_valve d ON a.node_id::integer=d.node_id::integer 
+					JOIN temp_anlgraf e ON a.node_id::integer=e.node_1::integer 
+					JOIN config_mincut_valve v ON v.id = c.id
+					WHERE closed=TRUE)
 					OR node_2::integer IN ('||v_text||' UNION
 					SELECT (a.node_id::integer) FROM node a JOIN cat_node b ON nodecat_id=b.id JOIN cat_feature_node c ON c.id=b.nodetype_id 
-					LEFT JOIN man_valve d ON a.node_id::integer=d.node_id::integer JOIN temp_anlgraf e ON a.node_id::integer=e.node_1::integer WHERE closed=TRUE)';
-			
+					LEFT JOIN man_valve d ON a.node_id::integer=d.node_id::integer 
+					JOIN temp_anlgraf e ON a.node_id::integer=e.node_1::integer 
+					JOIN config_mincut_valve v ON v.id = c.id
+					WHERE closed=TRUE)';
 			EXECUTE v_querytext;
 
 			v_text =  concat ('SELECT * FROM (',v_text,')a JOIN temp_anlgraf e ON a.node_id::integer=e.node_1::integer');
 
+			-- close customized stoppers acording on grafconfig column on mapzone table
+			EXECUTE 'UPDATE temp_anlgraf SET flag = 1 WHERE node_1 IN (SELECT (json_array_elements_text((grafconfig->>''stopper'')::json)) as node_id FROM '||quote_ident(v_table)||')';
+			EXECUTE 'UPDATE temp_anlgraf SET flag = 1 WHERE node_2 IN (SELECT (json_array_elements_text((grafconfig->>''stopper'')::json)) as node_id FROM '||quote_ident(v_table)||')';
+
+				
 			-- open boundary conditions set flag=0 for graf delimiters that have been setted to 1 on query before BUT ONLY ENABLING the right sense (to_arc)
 			
 			-- in function of graf class
@@ -366,6 +398,16 @@ BEGIN
 					where grafconfig is not null order by 1,2) a
 					ON to_arc::integer=arc_id::integer WHERE node_id::integer=node_1::integer);		
 			END IF;
+
+
+			-- close custom nodes acording init parameters
+			UPDATE temp_anlgraf SET flag = 1 WHERE node_1 IN (SELECT json_array_elements_text((v_parameters->>'forceClosed')::json));
+			UPDATE temp_anlgraf SET flag = 1 WHERE node_2 IN (SELECT json_array_elements_text((v_parameters->>'forceClosed')::json));
+
+			-- open custom nodes acording init parameters
+			UPDATE temp_anlgraf SET flag = 0 WHERE node_1 IN (SELECT json_array_elements_text((v_parameters->>'forceOpen')::json));
+			UPDATE temp_anlgraf SET flag = 0 WHERE node_2 IN (SELECT json_array_elements_text((v_parameters->>'forceOpen')::json));
+
 						
 			IF v_debug IS NULL OR v_debug IS FALSE THEN
 
@@ -624,6 +666,14 @@ BEGIN
 			END IF;
 		END IF;
 	END IF;
+
+	-- restore state selector (if it's needed)
+	IF v_usepsector IS NOT TRUE THEN
+		INSERT INTO selector_psector (psector_id, cur_user)
+		select unnest(text_column::integer[]), current_user from temp_table where fid=199 and cur_user=current_user
+		ON CONFLICT (psector_id, cur_user) DO NOTHING;
+	END IF;
+	
 	-- insert spacers on log
 	INSERT INTO audit_check_data (fid,  criticity, error_message) VALUES (v_fid,  3, '');
 	INSERT INTO audit_check_data (fid,  criticity, error_message) VALUES (v_fid,  2, '');
@@ -730,7 +780,7 @@ BEGIN
 	v_result_line := COALESCE(v_result_line, '{}'); 
 
 	--  Return
-	RETURN  ('{"status":"'||v_status||'", "message":{"level":'||v_level||', "text":"'||v_message||'"}, "version":"'||v_version||'"'||
+	RETURN ('{"status":"'||v_status||'", "message":{"level":'||v_level||', "text":"'||v_message||'"}, "version":"'||v_version||'"'||
              ',"body":{"form":{}, "data":{ "info":'||v_result_info||','||
 					  '"setVisibleLayers":["'||v_visible_layer||'"],'||
   					  '"setStyle":"Mapzones",'||
